Merge theme with "editor.tokenColorCustomizations"
Technically it only merges with `"editor.tokenColorCustomizations.textMateRules"` but that's a good start. 
It only merges if the override has a value set so we don't remove anything if a value isn't set. 
There's probably a better approach to
```typescript
if (override.foreground) {
        mergedRule.foreground = override.foreground
    }
```

I would have liked something like an `Optional` or similar. 
```typescript
override.foreground.map(mergedRule.foreground.assign)
```